Starting to hack out assimp specific stuff
