+++ Cargo.toml
@@ -10,10 +10,11 @@ colabrodo_common = { git = 'https://github.com/InsightCenterNoodles/colabrodo', 
 env_logger = "0.10.0"
 log = "0.4"
 notify = { version = "5.1", default-features = false, features = ["macos_kqueue"] }
-russimp = { version = "2.0", features = ["static-link"] }
+russimp = { version = "2.0", features = ["static-link"], optional=true }
 nalgebra-glm = "0.18"
 closure = "0.3"
 num-traits = "0.2.15"
+gltf = "1.1"
 
 [dependencies.uuid]
 version = "1.3"
@@ -22,3 +23,7 @@ features = [
     "fast-rng",          
     "macro-diagnostics", 
 ]
+
+
+[features]
+use_assimp = ["dep:russimp"]
